+++ sdk/storage/azure-storage-blob-data-movement/src/main/java/module-info.java
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-module com.azure.storage.blob.data.movement {
-    requires transitive com.azure.storage.common.data.movement;
-
-    exports com.azure.storage.blob.data.movement;
-}