+++ sdk/storage/azure-storage-blob-data-movement/src/test/java/com/azure/storage/blob/data/movement/APISpec.groovy
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-package com.azure.storage.blob.data.movement
-
-import com.azure.storage.common.test.shared.StorageSpec
-
-class APISpec extends StorageSpec {
-}