+++ sdk/storage/azure-storage-common-data-movement/src/main/java/com/azure/storage/common/data/movement/package-info.java
@@ -1,7 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/**
- * Package containing the common classes for Azure Storage Data movement.
- */
-package com.azure.storage.common.data.movement;