+++ sdk/storage/azure-storage-data-movement/src/main/java/com/azure/storage/data/movement/package-info.java
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+/**
+ * Package containing the classes for Azure Storage Data movement.
+ */
+package com.azure.storage.data.movement;