+++ sdk/storage/azure-storage-data-movement/CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## 12.0.0-beta.1 (Unreleased)
 
-### Features included in `azure-storage-blob-data-movement`
-- This package supports data movement for blob storage.
+### Features included in `azure-storage-data-movement`
+- This package supports functionality for storage data movement library.
 - Reactive streams support using [Project Reactor](https://projectreactor.io/).